add test directory and PosTest

+++ src/Main/Screen.java
@@ -3,11 +3,11 @@ package Main;
 import java.util.ArrayList;
 
 public class Screen {
-    public static final double WIDTH = 800;
-    public static final double HEIGHT = 600;
+    public static final int WIDTH = 800;
+    public static final int HEIGHT = 600;
     public static final int INTERVALS = 1000;
-    public static final double xAxis = WIDTH / 2;
-    public static final double yAxis = HEIGHT / 2;
+    public static final double xAxis = (double) WIDTH / 2;
+    public static final double yAxis = (double) HEIGHT / 2;
     public static final Pos offPosOfAxis = new Pos(0, Screen.yAxis);
     private ArrayList<Objects> objects;
